requirements.txt
apache-airflow>2.2
google-cloud-bigquery>=1.24.0
google-cloud-storage>=2.0
